requirements.txt
Flask
pandas
scikit-learn
google-generativeai